Restructured some things after feedback
